# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API",
    "version": "0.0.0"
  },
...
  "paths": {
...
    "/api/v1/speakers/": {
      "get": {
        "tags": [
          "speakers"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "fullName",
                "createdAt"
              ],
              "default": "createdAt"
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "in": "query",
            "name": "order",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "fullName": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 120
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          },
                          "bio": {
                            "type": "string",
                            "maxLength": 2000,
                            "nullable": true
                          },
                          "company": {
                            "type": "string",
                            "maxLength": 120,
                            "nullable": true
                          },
                          "website": {
                            "type": "string",
                            "format": "uri",
                            "nullable": true
                          },
                          "isConfirmed": {
                            "type": "boolean"
                          },
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdBy": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          }
                        },
                        "required": [
                          "fullName",
                          "email",
                          "isConfirmed",
                          "id",
                          "createdAt",
                          "updatedAt",
                          "createdBy"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "page",
                        "pageSize",
                        "total",
                        "totalPages"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "items",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "speakers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 2000,
                    "nullable": true
                  },
                  "company": {
                    "type": "string",
                    "maxLength": 120,
                    "nullable": true
                  },
                  "website": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "isConfirmed": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "fullName",
                  "email",
                  "isConfirmed"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fullName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "bio": {
                      "type": "string",
                      "maxLength": 2000,
                      "nullable": true
                    },
                    "company": {
                      "type": "string",
                      "maxLength": 120,
                      "nullable": true
                    },
                    "website": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true
                    },
                    "isConfirmed": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    }
                  },
                  "required": [
                    "fullName",
                    "email",
                    "isConfirmed",
                    "id",
                    "createdAt",
                    "updatedAt",
                    "createdBy"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/speakers/{id}": {
      "get": {
        "tags": [
          "speakers"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fullName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "bio": {
                      "type": "string",
                      "maxLength": 2000,
                      "nullable": true
                    },
                    "company": {
                      "type": "string",
                      "maxLength": 120,
                      "nullable": true
                    },
                    "website": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true
                    },
                    "isConfirmed": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    }
                  },
                  "required": [
                    "fullName",
                    "email",
                    "isConfirmed",
                    "id",
                    "createdAt",
                    "updatedAt",
                    "createdBy"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "speakers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "fullName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 2000,
                    "nullable": true
                  },
                  "company": {
                    "type": "string",
                    "maxLength": 120,
                    "nullable": true
                  },
                  "website": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  },
                  "isConfirmed": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "fullName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "email": {
                      "type": "string",
                      "format": "email"
                    },
                    "bio": {
                      "type": "string",
                      "maxLength": 2000,
                      "nullable": true
                    },
                    "company": {
                      "type": "string",
                      "maxLength": 120,
                      "nullable": true
                    },
                    "website": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true
                    },
                    "isConfirmed": {
                      "type": "boolean"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    }
                  },
                  "required": [
                    "fullName",
                    "email",
                    "isConfirmed",
                    "id",
                    "createdAt",
                    "updatedAt",
                    "createdBy"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "speakers"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/talks/": {
      "get": {
        "tags": [
          "talks"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "product",
                "engineering",
                "design",
                "business"
              ]
            },
            "in": "query",
            "name": "track",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "intro",
                "intermediate",
                "advanced"
              ]
            },
            "in": "query",
            "name": "level",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "eventId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "roomId",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "title",
                "startsAt",
                "createdAt"
              ],
              "default": "createdAt"
            },
            "in": "query",
            "name": "sort",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "default": "desc"
            },
            "in": "query",
            "name": "order",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string",
                            "minLength": 1,
                            "maxLength": 120
                          },
                          "abstract": {
                            "type": "string",
                            "maxLength": 4000,
                            "nullable": true
                          },
                          "track": {
                            "type": "string",
                            "enum": [
                              "product",
                              "engineering",
                              "design",
                              "business"
                            ]
                          },
                          "level": {
                            "type": "string",
                            "enum": [
                              "intro",
                              "intermediate",
                              "advanced"
                            ]
                          },
                          "startsAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "endsAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "isRecorded": {
                            "type": "boolean"
                          },
                          "eventId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "roomId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "createdBy": {
                            "type": "string",
                            "format": "uuid",
                            "nullable": true
                          }
                        },
                        "required": [
                          "title",
                          "track",
                          "level",
                          "startsAt",
                          "endsAt",
                          "isRecorded",
                          "eventId",
                          "roomId",
                          "id",
                          "createdAt",
                          "updatedAt",
                          "createdBy"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "page": {
                          "type": "integer"
                        },
                        "pageSize": {
                          "type": "integer"
                        },
                        "total": {
                          "type": "integer"
                        },
                        "totalPages": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "page",
                        "pageSize",
                        "total",
                        "totalPages"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "items",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "talks"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "abstract": {
                    "type": "string",
                    "maxLength": 4000,
                    "nullable": true
                  },
                  "track": {
                    "type": "string",
                    "enum": [
                      "product",
                      "engineering",
                      "design",
                      "business"
                    ]
                  },
                  "level": {
                    "type": "string",
                    "enum": [
                      "intro",
                      "intermediate",
                      "advanced"
                    ]
                  },
                  "startsAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endsAt": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "isRecorded": {
                    "type": "boolean"
                  },
                  "eventId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "roomId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "title",
                  "track",
                  "level",
                  "startsAt",
                  "endsAt",
                  "isRecorded",
                  "eventId",
                  "roomId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 120
                    },
                    "abstract": {
                      "type": "string",
                      "maxLength": 4000,
                      "nullable": true
                    },
                    "track": {
                      "type": "string",
                      "enum": [
                        "product",
                        "engineering",
                        "design",
                        "business"
                      ]
                    },
                    "level": {
                      "type": "string",
                      "enum": [
                        "intro",
                        "intermediate",
                        "advanced"
                      ]
                    },
                    "startsAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "endsAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "isRecorded": {
                      "type": "boolean"
                    },
                    "eventId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "roomId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdBy": {
                      "type": "string",
                      "format": "uuid",
                      "nullable": true
                    }
                  },
                  "required": [
                    "title",
                    "track",
                    "level",
                    "startsAt",
                    "endsAt",
                    "isRecorded",
                    "eventId",
                    "roomId",
                    "id",
                    "createdAt",
                    "updatedAt",
                    "createdBy"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}